+++ packages/jsonapi/src/Properties/ToOneRelationshipConfig.php
@@ -7,9 +7,9 @@ namespace EDT\JsonApi\Properties;
 use EDT\JsonApi\ResourceTypes\ResourceTypeInterface;
 use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Wrapping\Contracts\Types\TransferableTypeInterface;
-use EDT\Wrapping\Contracts\Types\TypeInterface;
 use EDT\Wrapping\Properties\ToOneRelationshipUpdatability;
 use EDT\Wrapping\Properties\ToOneRelationshipReadability;
+use EDT\Wrapping\WrapperFactories\WrapperObject;
 
 /**
  * @template TCondition of PathsBasedInterface
@@ -39,7 +39,7 @@ class ToOneRelationshipConfig extends AbstractConfig
     }
 
     /**
-     * @param null|callable(TEntity): (TRelationship|null) $customReadCallback
+     * @param null|callable(TEntity): (TRelationship|WrapperObject<TRelationship>|null) $customReadCallback
      *
      * @return $this
      *